CI: el ultimo estrechamiento de entero del codigo
Windows volvio a caer, y otra vez por lo mismo: en test_moral.cpp el emisor
llegaba como uint32_t y RawCommand::emitter es uint16_t. gcc y clang callan;
MSVC avisa C4244 y con -Werror eso no rompe una prueba, rompe la compilacion
entera de la matriz de Windows. Cuarta vez.

Esta vez, en vez de arreglar el sitio y esperar el siguiente viaje a la CI,
barri el fichero con clang -Wconversion en local y aparecio uno mas — en
data_catalog.hpp, el `uint16 << 8` de u16() que PROMOCIONA a int y vuelve a
estrecharse al asignar. MSVC no lo habia marcado todavia, pero es la misma
trampa esperando a que la pise otra unidad de traduccion.

Los dos cerrados. Ya no queda ningun estrechamiento de entero en el codigo:
clang -Wconversion sale limpio.

Suite 54/54.

+++ tests/unit/test_moral.cpp
@@ -42,7 +42,11 @@ static void fill_spawn(RawCommand& c, uint32_t emitter, uint32_t seq,
                        uint32_t unit_class, uint32_t speed_mtpt) {
     std::memset(&c, 0, sizeof(RawCommand));
     c.target_tick  = 0;
-    c.emitter      = emitter;
+    // static_cast explicito: `emitter` es uint16_t en RawCommand y el parametro
+    // llega como uint32_t. MSVC avisa (C4244) donde gcc y clang callan, y con
+    // -Werror eso no rompe una prueba: rompe la CI entera de Windows. Ya nos
+    // costo una tarde en el sprint 1.43.
+    c.emitter      = static_cast<uint16_t>(emitter);
     c.type         = CommandType::SPAWN_UNIT;
     c.sequence     = seq;
     c.p.x_raw      = x_raw;
